requirements.txt
aiohttp
playwright
pandas
numpy
pyarrow
matplotlib
tldextract
charset-normalizer
pydantic
tqdm
rich
pyyaml